Server/requirements.txt
fastapi==0.115.8
ipython==8.12.3
Pillow==11.1.0
redis~=5.2.1
Requests==2.32.3
uvicorn==0.34.0
websockets~=15.0
google-auth
